[TC-42] - Exception 추가

+++ src/main/java/com/server/booyoungee/domain/BookMark/application/BookMarkService.java
@@ -5,6 +5,10 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import com.server.booyoungee.domain.BookMark.exception.DuplicateBookMarkException;
+import com.server.booyoungee.domain.BookMark.exception.NotFoundBookMarkException;
+import com.server.booyoungee.domain.place.exception.movie.NotFoundMoviePlaceException;
+import com.server.booyoungee.domain.place.exception.store.NotFoundStorePlaceException;
 import org.springframework.stereotype.Service;
 import org.webjars.NotFoundException;
 
@@ -18,6 +22,8 @@ import com.server.booyoungee.domain.user.domain.User;
 
 import lombok.RequiredArgsConstructor;
 
+import static com.server.booyoungee.domain.BookMark.exception.BookMarkExceptionCode.NOT_FOUND_BOOKMARK;
+
 @Service
 @RequiredArgsConstructor
 public class BookMarkService {
@@ -25,14 +31,14 @@ public class BookMarkService {
 	private final BookMarkRepository bookMarkRepository;
 	private final PlaceService placeService;
 
-	public Object getBookMarks(User user) throws IOException {
+	public List<TourInfoBookMarkDto> getBookMarks(User user) throws IOException {
 		List<BookMark> bookMarkList = bookMarkRepository.findAllByUser(user);
 		List<TourInfoBookMarkDto> dto = bookMarkList.stream()
 			.map(bookMark -> {
 				try {
-					return getPlace(bookMark.getPlaceId(), bookMark.getType());
+					return getPlace(bookMark.getBookMarkId(),bookMark.getPlaceId(), bookMark.getType());
 				} catch (IOException e) {
-					throw new NotFoundException("해당 북마크를 찾을 수 없습니다.");
+					throw new NotFoundBookMarkException();
 				}
 			})
 			.collect(Collectors.toList());
@@ -41,7 +47,7 @@ public class BookMarkService {
 
 	}
 
-	public Object getMyBookMarkDetails(User user) throws IOException {
+	public List<PlaceDetailsDto> getMyBookMarkDetails(User user) throws IOException {
 		List<BookMark> bookMarkList = bookMarkRepository.findAllByUser(user);
 		List<PlaceDetailsDto> dto = new ArrayList<>();
 		for (BookMark bookMark : bookMarkList) {
@@ -53,14 +59,24 @@ public class BookMarkService {
 
 	public void addBookMark(User user, Long placeId, PlaceType type) {
 		try {
-			placeService.getPlace(placeId, type);
+			TourInfoBookMarkDto dto = placeService.getPlace(1L,placeId, type);
 		} catch (Exception e) {
+			if(type.equals(PlaceType.movie))
+			{
+				throw new NotFoundMoviePlaceException();
+			}else if(type.equals(PlaceType.store))
+			{
+				throw new NotFoundStorePlaceException();
+			}
+			else{
+				throw new NotFoundException("contentID : " + placeId + " not found.");
+			}
 
 		}
-		System.out.println("id: " + placeId + "user: " + user.getUserId() + "type: " + type);
-		if (user == null || placeId == null || type == null) {
-			throw new NotFoundException("Movie place with ID " + placeId + " not found.");
+		if(bookMarkRepository.existsByUserIdAndPlaceIdAndType(user, placeId,type)){
+			throw new DuplicateBookMarkException();
 		}
+
 
 		BookMark bookMark = BookMark.builder()
 			.userId(user)
@@ -70,12 +86,15 @@ public class BookMarkService {
 		bookMarkRepository.save(bookMark);
 	}
 
-	public TourInfoBookMarkDto getPlace(Long placeId, PlaceType type) throws IOException {
-		return placeService.getPlace(placeId, type);
+	public TourInfoBookMarkDto getPlace(Long id,Long placeId, PlaceType type) throws IOException {
+		return placeService.getPlace(id,placeId, type);
 	}
 
 	public void deleteBookMark(User user, Long bookMarkId) {
 		BookMark bookMark = bookMarkRepository.findByBookMarkIdAndUserId(bookMarkId, user);
+		if (bookMark == null) {
+			throw new NotFoundBookMarkException();
+		}
 		bookMarkRepository.delete(bookMark);
 	}
